% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcturbo.R
\docType{package}
\name{dbcturbo-package}
\alias{dbcturbo}
\alias{dbcturbo-package}
\title{dbcturbo: High-Performance Streaming Reader for 'DATASUS' DBC Files}
\description{
A modern, memory-efficient R package for reading and converting DBC files produced by the Brazilian Ministry of Health's 'DATASUS' system. DBC files are DBF databases compressed with the 'PKWare' implode algorithm. Unlike 'read.dbc', this package processes files in streaming mode, never loading the full dataset into RAM. It can export directly to CSV or DBF and provides R helpers for in-memory reading and Parquet conversion.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GPimentel14/dbcturbo}
  \item Report bugs at \url{https://github.com/GPimentel14/dbcturbo/issues}
}

}
\author{
\strong{Maintainer}: Gumercindo Pimentel Peralta \email{gumercindopimentel@gmail.com} (\href{https://orcid.org/0000-0002-5896-3171}{ORCID})

Authors:
\itemize{
  \item Gumercindo Pimentel Peralta \email{gumercindopimentel@gmail.com} (\href{https://orcid.org/0000-0002-5896-3171}{ORCID})
}

Other contributors:
\itemize{
  \item Juliana da Silva (\href{https://orcid.org/0000-0002-1089-6766}{ORCID}) [thesis advisor]
  \item Mark Adler (Author of blast.c (PKWare implode decompressor)) [contributor]
  \item Daniela Petruzalek (Author of read.dbc, dbc2dbf.c) [contributor]
  \item Pablo Fonseca (Author of blast-dbf) [contributor]
}

}
\keyword{internal}
